z3str3: refactoring, move regex automata methods to theory_str_regex
