+++ src/smt/CMakeLists.txt
@@ -67,6 +67,7 @@ z3_add_component(smt
     theory_seq.cpp
     theory_special_relations.cpp
     theory_str.cpp
+    theory_str_regex.cpp
     theory_utvpi.cpp
     theory_wmaxsat.cpp
     uses_theory.cpp
